Widen input without a float dtype to float64
Integer arrays, lists and Series carry no dtype the caller asked for, and
they were being coerced to float32, which is lossy from 2**24 onwards:

    rolling_mean(np.array([16777217, 16777219, 16777221]), 1, 1)
    # [16777216., 16777220., 16777220.]

For the free functions that came from the registration order, since pybind's
conversion pass takes the first overload registered; an exact dtype match is
resolved in the earlier no-conversion pass, so float32 and float64 arrays are
unaffected either way. The GroupedArray factory casts explicitly instead of
going through overload resolution, so it gets the same change directly.

This is a user-visible change: unspecified-dtype input now returns float64
and takes twice the memory. Callers who want float32 pass a float32 array,
which was already the only way to be sure of it.

+++ src/grouped_array.cpp
@@ -833,7 +833,7 @@ void init_ga(py::module_ &m) {
         }
         data = py::array::ensure(data, py::array::c_style);
         if (data.dtype().kind() != 'f') {
-          data = data.attr("astype")("float32");
+          data = data.attr("astype")("float64");
         }
         if (py::isinstance<py::array_t<float>>(data)) {
           return py::cast(std::make_unique<GroupedArray<float>>(
